Revert "DPFL-2305 (WIP) FOR TESTING 20240517"
