+++ ccd-definition/AuthorisationCaseField/CareSupervision/la-barrister.json
@@ -324,13 +324,6 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "CaseFieldID": "skipDraftOrderUrgencyPage",
-    "UserRole": "[LABARRISTER]",
-    "CRUD": "CRU"
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "CARE_SUPERVISION_EPO",
     "CaseFieldID": "c2ApplicationType",
     "UserRole": "[LABARRISTER]",
     "CRUD": "CRU"
