+++ ccd-definition/AuthorisationCaseField/CareSupervision/la-managing.json
@@ -646,13 +646,6 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "CaseFieldID": "skipDraftOrderUrgencyPage",
-    "UserRole": "[LAMANAGING]",
-    "CRUD": "CRU"
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "CARE_SUPERVISION_EPO",
     "CaseFieldID": "c2ApplicationType",
     "UserRole": "[LAMANAGING]",
     "CRUD": "CRU"
